+++ src/mesh/mesh_voxelizer_refine.cu
@@ -23,8 +23,8 @@ template <const ArgsPack *AP>
 __global__
 void Cu_MarkBlocks_MarkBoundary
 (
-    const int id_max_curr,
-    const int n_maxcblocks,
+    const long int id_max_curr,
+    const long int n_maxcblocks,
     const bool hit_max,
     const int L,
     int *__restrict__ cblock_ID_ref,
@@ -124,8 +124,8 @@ template <const ArgsPack *AP>
 __global__
 void Cu_MarkBlocks_MarkExterior
 (
-    const int id_max_curr,
-    const int n_maxcblocks,
+    const long int id_max_curr,
+    const long int n_maxcblocks,
     const bool hit_max,
     const int L,
     int *__restrict__ cblock_ID_ref,
@@ -222,8 +222,8 @@ template <const ArgsPack *AP>
 __global__
 void Cu_MarkBlocks_UpdateSolidChildren
 (
-    const int id_max_curr,
-    const int n_maxcblocks,
+    const long int id_max_curr,
+    const long int n_maxcblocks,
     const int L,
     const int *__restrict__ cblock_ID_mask,
     int *__restrict__ cblock_ID_nbr,
@@ -278,8 +278,8 @@ template <const ArgsPack *AP>
 __global__
 void Cu_MarkBlocks_Propagate
 (
-    const int id_max_curr,
-    const int n_maxcblocks,
+    const long int id_max_curr,
+    const long int n_maxcblocks,
     const int L,
     int *__restrict__ cblock_ID_ref,
     const int *__restrict__ cblock_ID_mask,
@@ -396,7 +396,7 @@ template <const ArgsPack *AP>
 __global__
 void Cu_MarkBlocks_Finalize
 (
-    const int id_max_curr,
+    const long int id_max_curr,
     int *__restrict__ cblock_ID_ref,
     int *__restrict__ tmp_1
 )
